refactor: make Content comp module
